Move downloading of facilities data to CSV file button to facility dashboard page
