+++ src/adminfacilitydashboard/styles.module.css
@@ -189,3 +189,25 @@
   letter-spacing: -0.5px;
   color: #6c5ce7;
 }
+
+.downloadBtn {
+  composes: primaryBtn from '/src/globalStyles.css';
+  width: clamp(20vw, 10em, 25vh);
+  margin: 10px auto 20px;
+  -moz-box-sizing: border-box;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  text-align: center;
+  text-decoration: none;
+  border: 1px solid #6C5CE7;
+}
+
+.downloadBtn:disabled, .downloadBtn[disabled] {
+  border: 1px solid #a59de1;
+  background: #a59de1;
+}
+
+.downloadBtn > span {
+  color: white;
+}
